[SPARK-56853][SQL][TESTS][FOLLOWUP] Address review feedback
Follow-up changes after the initial PR landed on the branch:

- Strip the SPARK-56853 JIRA tag from test names, comments, and file
  headers across the new tests. These are feature-coverage tests, not
  bug regressions, so the JIRA label does not belong inline.
- Correct the inaccurate "view's home schema" framing in the two new
  PATH-disabled compat tests in `SQLViewSuite`. The fallback target is
  the view's `viewCatalogAndNamespace` property -- i.e. the creator
  session's USE state captured at CREATE VIEW time -- not the schema
  the view physically lives in. The previous comment overgeneralized
  from a setup choice (USE-ing the destination schema before CREATE)
  to the general behavior.
- Merge `PathElementSuite` into `CatalogManagerSuite`. The seven
  `PathElement.validateNoStaticDuplicates` tests are pure data tests
  living in the same package as the consumer; folding them into the
  sibling suite keeps the test file count down without losing coverage.
  `CatalogManagerSuite` is now 16 tests grouped by section comment.
- Restructure `sql-tests/inputs/sql-path.sql` as the primary SQL-level
  reference for the feature. Adds a self-describing header, a Table of
  Contents, and ten clearly delimited sections covering: default path
  observability; SET PATH grammar (literal, DEFAULT_PATH, SYSTEM_PATH,
  PATH append, current_schema/current_database); CURRENT_PATH() builtin
  and its ANSI no-parens form; the full set of static error conditions;
  routine resolution via PATH for scalar AND table functions; relation
  resolution via PATH; persisted view frozen-path behavior including
  invoker-context current_schema/current_path; SQL function frozen-path
  behavior for scalar and table functions including the invoker-context
  guard; the `spark.sql.defaultPath` conf (explicit override, expansion,
  rejection of invalid values and the PATH keyword); and PATH-disabled
  behaviors. The corresponding `results/` and `analyzer-results/`
  golden files were regenerated and reviewed.
